pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lmma_trendminer"
version = "0.1.0"
description = "Fast vector-first trend analysis (vector-only retrieval → UMAP → HDBSCAN → minimal text fetch)."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Kuppuram Govindan" }]
keywords = ["nlp", "trend analysis", "clustering", "umap", "hdbscan", "vector search", "opensearch"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Intended Audience :: Developers",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Text Processing :: Linguistic",
  "Operating System :: OS Independent"
]

# dependencies = [
#  "numpy>=1.24",
#  "scikit-learn>=1.3",
#  "umap-learn>=0.5",
#  "hdbscan>=0.8",
#  "transformers>=4.44",
#  "torch>=2.1",
#  "scipy>=1.10",
#  "opensearch-py>=2.5",
#  "dateparser>=1.2",   # ← add this
#  "tzdata>=2024.1",    # ← good to have on Windows for timezones
#]

dependencies = [
  "numpy>=1.24",
  "scikit-learn>=1.3",
  "umap-learn>=0.5",
  "hdbscan>=0.8",
  "transformers>=4.44",
  "torch>=2.1",
  "scipy>=1.10",
  "opensearch-py>=2.5",
  "pyyaml>=6.0.1",         # <-- add
]

[project.optional-dependencies]
demo = ["fastapi>=0.111", "uvicorn>=0.29"]
vectorstores = ["weaviate-client>=4.5", "pinecone-client>=4.0", "qdrant-client>=1.8"]
ingest = [
  "pandas>=2.0",
  "opensearch-py>=2.5",
  "tqdm>=4.66",
  "pyyaml>=6.0.1",
  "sentence-transformers>=2.5",
  "typer>=0.12.0"
]

train = [
  "transformers>=4.44,<5",
  "accelerate>=0.34",
  "datasets>=2.19",
  "seqeval>=1.2.2",
  "typer>=0.12.0",
  "pyyaml>=6.0.1"
  # NOTE: install PyTorch separately for your platform/CUDA, e.g.
  # pip install torch --index-url https://download.pytorch.org/whl/cu121
]

llm = [
  "groq>=0.9.0",
  "python-dotenv>=1.0.1"
]

dev = [
  "black>=24.3.0",
  "ruff>=0.5.0",
  "pytest>=8.0.0"
]

# optional: both
dev-all = [
  "lmma_trendminer[ingest,train,llm]"
]

[project.urls]
Homepage = "https://github.com/kuppuram/lmma_trendminer"
Issues = "https://github.com/kuppuram/lmma_trendminer/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["lmma_trendminer*"]
exclude = ["tests*", "examples*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
exclude = ["examples", "tests"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[project.scripts]
lmma-ingest = "tools.ingest.cli:app"     # Typer/Click entrypoint
lmma-train-ner = "tools.train.ner_cli:app"